Move reducer to sources
